multiple list support added

+++ src/lib/List.svelte
@@ -1,11 +1,31 @@
 <script lang="ts">
   import Item from "./Item.svelte";
+  import SideBar from "./SideBar.svelte";
   import { invoke } from "@tauri-apps/api/tauri";
 
-  let list_name = "list1";
+  let list_names = [];
+  let list_name = "";
+  let items = [];
   let item_name = "";
 
-  $: items = [];
+  async function load_list_names() {
+    let list = await invoke("get_list_names");
+    if (list) {
+      list_names = list;
+    }
+  }
+
+  async function deleteList() {
+    let response = await invoke("delete_list", {
+      listName: this.previousElementSibling.innerText,
+    });
+    if (response) {
+      list_names = response;
+      if (list_name == this.previousElementSibling.innerText) {
+        list_name = "";
+      }
+    }
+  }
 
   async function addItem() {
     if (item_name != "") {
@@ -15,13 +35,15 @@
       });
       if (list) {
         items = list;
+        item_name = "";
       }
     }
   }
 
   async function loadData() {
-    let list = await invoke("get_list", { listName: list_name });
+    let list = await invoke("get_list", { listName: this.innerText });
     if (list) {
+      list_name = this.innerText;
       items = list;
     }
   }
@@ -39,34 +61,37 @@
   async function deleteItem() {
     let list = await invoke("remove_item_from_list", {
       listName: list_name,
-      itemName: this.id,
+      itemName: this.previousElementSibling.innerText,
     });
     if (list) {
       items = list;
     }
   }
-  loadData();
+  load_list_names();
 </script>
 
 <div class="container">
-  <div>
-    <form on:submit|preventDefault={addItem}>
-      <input placeholder="Enter an item to list" bind:value={item_name} />
-      <button type="submit">Add</button>
-    </form>
-    <div class="list">
-      {#each items as item}
-        {#key item.name}
-          <Item
-            id={item.name}
-            name={item.name}
-            checked={item.checked}
-            onClick={toggleChecked}
-            onDelete={deleteItem}
-          />
-        {/key}
-      {/each}
-    </div>
+  <SideBar {list_names} onClick={loadData} onDelete={deleteList} />
+  <div class="right">
+    {#if list_name != ""}
+      <h3>{list_name}</h3>
+      <form on:submit|preventDefault={addItem}>
+        <input placeholder="Enter an item to list" bind:value={item_name} />
+        <button type="submit">Add</button>
+      </form>
+      <div class="list">
+        {#each items as item}
+          {#key item.name}
+            <Item
+              name={item.name}
+              checked={item.checked}
+              onClick={toggleChecked}
+              onDelete={deleteItem}
+            />
+          {/key}
+        {/each}
+      </div>
+    {/if}
   </div>
 </div>
 
@@ -75,11 +100,14 @@
     display: flex;
     flex-direction: row;
     align-items: flex-start;
-    justify-content: space-around;
+    justify-content: space-between;
     width: 100%;
   }
+  .right {
+    width: 70vw;
+  }
   .list {
-    padding: 10px;
+    margin-top: 10px;
     display: flex;
     flex-direction: column;
     align-items: center;
